+++ www/src/components/evaluation-table.js
@@ -1,11 +1,11 @@
 import React, { Component } from "react"
-import presets, { colors } from "../utils/presets"
+import presets, { colors, space, breakpoints } from "../utils/presets"
 import EvaluationCell from "./evaluation-cell"
 import infoIcon from "../assets/info-icon.svg"
 import SectionTitle from "./evaluation-table-section-title"
 import SectionHeaderTop from "./evaluation-table-section-header-top"
 import SectionHeaderBottom from "./evaluation-table-section-header-bottom"
-import { options, rhythm } from "../utils/typography"
+import { rhythm } from "../utils/typography"
 
 class EvaluationTable extends Component {
   constructor() {
@@ -18,16 +18,15 @@ class EvaluationTable extends Component {
       return [
         words.slice(0, words.length - 1).join(` `),
         <span
-          css={{
-            // WebkitHyphens: `auto`,
-            // MsHyphens: `auto`,
-            // hyphens: `auto`,
-            // wordBreak: `break-all`,
-            // display: `inline-block`,
-            "&:hover": {
-              background: colors.ui.bright,
-            },
-          }}
+          css={
+            {
+              // WebkitHyphens: `auto`,
+              // MsHyphens: `auto`,
+              // hyphens: `auto`,
+              // wordBreak: `break-all`,
+              // display: `inline-block`,
+            }
+          }
           key={`info-icon-${words[words.length - 1]}`}
         >
           {` `}
@@ -35,9 +34,9 @@ class EvaluationTable extends Component {
           <img
             src={infoIcon}
             css={{
-              height: rhythm(2 / 5),
-              marginBottom: rhythm(2 / 15),
-              verticalAlign: `text-bottom`,
+              height: rhythm(space[3]),
+              marginBottom: rhythm(space[1]),
+              verticalAlign: `baseline`,
             }}
             alt={`Info Icon`}
           />
@@ -56,30 +55,20 @@ class EvaluationTable extends Component {
                 display: `inline-block`,
                 marginLeft: `auto`,
                 marginRight: `auto`,
-                padding: `${rhythm(1 / 4)} 0 ${rhythm(1 / 4)} ${rhythm(1 / 4)}`,
-                [presets.Mobile]: {
-                  padding: `${rhythm(1 / 2)} 0 ${rhythm(1 / 2)} ${rhythm(
-                    1 / 2
-                  )}`,
-                },
               }}
             >
-              {/* eslint-disable jsx-a11y/anchor-is-valid */}
-              {/* jsx-a11y really wants us to change this to a button */}
-              {/* eslint-disable-next-line */}
-              <a
+              <button
                 css={{
-                  "&&": {
-                    fontWeight: `normal`,
-                    borderBottom: 0,
-                  },
+                  border: 0,
+                  padding: 0,
+                  textAlign: `left`,
                 }}
                 onClick={e => {
                   e.preventDefault()
                 }}
               >
                 {renderText(text)}
-              </a>
+              </button>
               {/* eslint-enable */}
             </div>
           )
@@ -127,9 +116,9 @@ class EvaluationTable extends Component {
                       <SectionHeaderBottom
                         display={row.node.Subcategory}
                         category={row.node.Subcategory}
-                        key={`section-header-bottom-${i}`}
+                        key={`section-header-${s}-bottom-${i}`}
                       />,
-                      <tr key={`first-row-${i}`}>
+                      <tr key={`section-${s}-first-row-${i}`}>
                         {headers.map((header, j) => (
                           <td
                             key={j}
@@ -146,15 +135,15 @@ class EvaluationTable extends Component {
                               paddingLeft: 0,
                               textAlign: `left`,
                               verticalAlign: `middle`,
-                              fontSize: `90%`,
-                              lineHeight: `${rhythm(3 / 4)}`,
+                              fontSize: presets.scale[1],
+                              lineHeight: presets.lineHeights.solid,
                             }}
                             id={
                               j === 0
                                 ? row.node.Feature.toLowerCase()
                                     .split(` `)
                                     .join(`-`)
-                                : false
+                                : undefined
                             }
                             onClick={() => {
                               this.setState({
@@ -170,22 +159,15 @@ class EvaluationTable extends Component {
                         style={{
                           display: showTooltip(s, i) ? `table-row` : `none`,
                         }}
-                        key={`second-row-${i}`}
+                        key={`section-${s}-second-row-${i}`}
                       >
                         <td
                           css={{
-                            fontFamily: options.headerFontFamily.join(`,`),
-                            paddingBottom: `calc(${rhythm(1)} - 1px)`,
+                            paddingBottom: `calc(${rhythm(space[5])} - 1px)`,
                             "&&": {
-                              paddingRight: `${rhythm(1 / 4)}`,
-                              paddingLeft: `${rhythm(1 / 4)}`,
-                              [presets.Mobile]: {
-                                paddingRight: `${rhythm(1 / 2)}`,
-                                paddingLeft: `${rhythm(1 / 2)}`,
-                              },
-                              [presets.Phablet]: {
-                                paddingRight: `${rhythm(2)}`,
-                                paddingLeft: `${rhythm(2)}`,
+                              [breakpoints.xs]: {
+                                paddingRight: `${rhythm(space[3])}`,
+                                paddingLeft: `${rhythm(space[3])}`,
                               },
                             },
                           }}