+++ www/src/components/showcase-details.js
@@ -4,8 +4,14 @@ import url from "url"
 import Img from "gatsby-image"
 import qs from "qs"
 
-import presets, { colors } from "../utils/presets"
-import { options, scale, rhythm } from "../utils/typography"
+import presets, {
+  colors,
+  space,
+  radii,
+  shadows,
+  breakpoints,
+} from "../utils/presets"
+import { options, rhythm } from "../utils/typography"
 import sharedStyles from "../views/shared/styles"
 import { Link, StaticQuery, graphql } from "gatsby"
 import Layout from "../components/layout"
@@ -17,8 +23,8 @@ import FeaturedIcon from "../assets/featured-detailpage-featuredicon.svg"
 import FeatherIcon from "../assets/showcase-feather.svg"
 import GithubIcon from "react-icons/lib/go/mark-github"
 
-const gutter = rhythm(3 / 4)
-const gutterDesktop = rhythm(6 / 4)
+const gutter = rhythm(space[6])
+const gutterDesktop = rhythm(space[8])
 
 const styles = {
   link: {
@@ -32,17 +38,14 @@ const styles = {
     top: 280,
     width: 300,
     transform: `translateX(-75px) rotate(90deg)`,
-    [presets.Desktop]: {
-      ...scale(-1 / 6),
-    },
   },
   prevNextLinkSiteTitle: {
     color: colors.gatsby,
     fontWeight: `bold`,
   },
   prevNextImage: {
-    borderRadius: presets.radius,
-    boxShadow: `0 0 38px -8px ${colors.gatsby}`,
+    borderRadius: radii[1],
+    boxShadow: shadows.overlay,
   },
   prevNextPermalinkLabel: {
     color: colors.gray.calm,
@@ -51,7 +54,7 @@ const styles = {
   },
   prevNextPermalinkImage: {
     marginBottom: 0,
-    marginTop: rhythm(options.blockMarginBottom),
+    marginTop: rhythm(space[6]),
   },
   prevNextPermalinkTitle: {
     color: colors.gatsby,
@@ -73,12 +76,12 @@ const styles = {
     marginRight: 4,
     verticalAlign: `sub`,
     position: `absolute`,
-    left: `-${rhythm(3 / 4)}`,
+    left: `-${rhythm(space[6])}`,
     top: `50%`,
     transform: `translateY(-50%)`,
   },
   prevNextPermalinkMeta: {
-    marginLeft: rhythm(6 / 4),
+    marginLeft: rhythm(space[8]),
     display: `flex`,
     flexDirection: `row`,
     justifyContent: `flex-end`,
@@ -159,11 +162,7 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                 transform: `translateX(750px)`,
               }}
             >
-              <div
-                css={{
-                  margin: `25px`,
-                }}
-              >
+              <div css={{ margin: `25px` }}>
                 <Img
                   key={nextSite.id}
                   css={{
@@ -181,20 +180,10 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                   alt=""
                 />
               </div>
-              <div
-                css={{
-                  ...styles.prevNextLink,
-                }}
-              >
+              <div css={styles.prevNextLink}>
                 <MdArrowUpward />
                 <div> Next Site in Showcase </div>
-                <div
-                  css={{
-                    ...styles.prevNextLinkSiteTitle,
-                  }}
-                >
-                  {nextSite.title}
-                </div>
+                <div css={styles.prevNextLinkSiteTitle}>{nextSite.title}</div>
               </div>
             </Link>
           }
@@ -212,11 +201,7 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                 transform: `translateX(-100%)`,
               }}
             >
-              <div
-                css={{
-                  margin: `25px`,
-                }}
-              >
+              <div css={{ margin: `25px` }}>
                 <Img
                   key={previousSite.id}
                   css={{
@@ -243,11 +228,7 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
               >
                 <MdArrowUpward />
                 <div> Previous Site in Showcase </div>
-                <div
-                  css={{
-                    ...styles.prevNextLinkSiteTitle,
-                  }}
-                >
+                <div css={styles.prevNextLinkSiteTitle}>
                   {previousSite.title}
                 </div>
               </div>
@@ -287,43 +268,26 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
               </Helmet>
               <div
                 css={{
-                  fontFamily: options.headerFontFamily.join(`,`),
                   padding: gutter,
-                  paddingBottom: rhythm(1.5 / 4),
-                  [presets.Desktop]: {
+                  paddingBottom: rhythm(space[2]),
+                  [breakpoints.lg]: {
                     padding: gutterDesktop,
-                    paddingBottom: rhythm(3 / 4),
+                    paddingBottom: rhythm(space[6]),
                   },
                 }}
               >
-                <h1
-                  css={{
-                    margin: 0,
-                  }}
-                >
-                  {data.sitesYaml.title}
-                </h1>
+                <h1 css={{ margin: 0 }}>{data.sitesYaml.title}</h1>
                 <a
                   href={data.sitesYaml.main_url}
                   css={{
                     ...styles.link,
                     fontWeight: `bold`,
-                    [presets.Desktop]: {
-                      ...scale(-1 / 6),
-                    },
                   }}
                 >
                   {cleanUrl(data.sitesYaml.main_url)}
                 </a>
                 {data.sitesYaml.built_by && (
-                  <span
-                    css={{
-                      color: colors.gray.calm,
-                      [presets.Desktop]: {
-                        ...scale(-1 / 6),
-                      },
-                    }}
-                  >
+                  <span css={{ color: colors.gray.calm }}>
                     <span
                       css={{
                         paddingRight: 8,
@@ -355,7 +319,7 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                   borderTop: `1px solid ${colors.ui.light}`,
                   fontFamily: options.headerFontFamily.join(`,`),
                   margin: `0 ${gutter}`,
-                  [presets.Desktop]: {
+                  [breakpoints.lg]: {
                     margin: `0 ${gutterDesktop}`,
                   },
                 }}
@@ -369,9 +333,6 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                       fontWeight: `bold`,
                       padding: 20,
                       paddingLeft: 0,
-                      [presets.Desktop]: {
-                        ...scale(-1 / 6),
-                      },
                     }}
                   >
                     <img
@@ -392,9 +353,6 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                       paddingLeft: data.sitesYaml.featured ? false : 0,
                       display: `flex`,
                       borderRight: `1px solid ${colors.ui.light}`,
-                      [presets.Desktop]: {
-                        ...scale(-1 / 6),
-                      },
                     }}
                   >
                     <a
@@ -405,7 +363,7 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                     >
                       <GithubIcon
                         style={{
-                          fontSize: 20,
+                          fontSize: presets.scale[4],
                           marginRight: 10,
                           verticalAlign: `text-bottom`,
                         }}
@@ -449,26 +407,21 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                     <a
                       href={data.sitesYaml.main_url}
                       css={{
+                        backgroundColor: colors.gatsby,
                         border: 0,
-                        borderRadius: presets.radius,
-                        color: colors.gatsby,
+                        borderRadius: radii[1],
                         display: shouldShowVisitButtonOnMobile ? `none` : null,
                         fontFamily: options.headerFontFamily.join(`,`),
                         fontWeight: `bold`,
-                        marginRight: rhythm(1.5 / 4),
-                        padding: `${rhythm(1 / 5)} ${rhythm(2 / 3)}`,
+                        marginRight: rhythm(space[2]),
+                        padding: `${rhythm(space[1])} ${rhythm(space[4])}`,
                         textDecoration: `none`,
                         WebkitFontSmoothing: `antialiased`,
                         "&&": {
-                          backgroundColor: colors.gatsby,
+                          color: colors.white,
                           borderBottom: `none`,
-                          boxShadow: `none`,
-                          color: `white`,
-                          "&:hover": {
-                            backgroundColor: colors.gatsby,
-                          },
                         },
-                        [shouldShowVisitButtonOnMobile && presets.Phablet]: {
+                        [shouldShowVisitButtonOnMobile && breakpoints.sm]: {
                           display: `block`,
                         },
                       }}
@@ -493,11 +446,7 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
                   </div>
                 </div>
               </div>
-              <div
-                css={{
-                  borderTop: `1px solid ${colors.ui.light}`,
-                }}
-              >
+              <div css={{ borderTop: `1px solid ${colors.ui.light}` }}>
                 <Img
                   key={data.sitesYaml.id}
                   fluid={
@@ -515,7 +464,7 @@ const ShowcaseDetails = ({ parent, data, isModal, categories }) => (
               <div
                 css={{
                   padding: gutter,
-                  [presets.Desktop]: {
+                  [breakpoints.lg]: {
                     padding: gutterDesktop,
                   },
                 }}