+++ www/src/templates/tags.js
@@ -6,6 +6,7 @@ import BlogPostPreviewItem from "../components/blog-post-preview-item"
 import Button from "../components/button"
 import Container from "../components/container"
 import Layout from "../components/layout"
+import { space } from "../utils/presets"
 import { rhythm } from "../utils/typography"
 
 const Tags = ({ pageContext, data, location }) => {
@@ -19,14 +20,17 @@ const Tags = ({ pageContext, data, location }) => {
     <Layout location={location}>
       <Container>
         <h1>{tagHeader}</h1>
-        <Button tiny key="blog-post-view-all-tags-button" to="/blog/tags">
+        <Button small key="blog-post-view-all-tags-button" to="/blog/tags">
           View All Tags <TagsIcon />
         </Button>
         {edges.map(({ node }) => (
           <BlogPostPreviewItem
             post={node}
             key={node.fields.slug}
-            css={{ marginBottom: rhythm(2) }}
+            css={{
+              marginTop: rhythm(space[9]),
+              marginBottom: rhythm(space[9]),
+            }}
           />
         ))}
       </Container>