+++ packages/gatsby-plugin-cxs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-cxs/compare/gatsby-plugin-cxs@2.0.7...gatsby-plugin-cxs@2.0.8) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-plugin-cxs
+
+## [2.0.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-cxs/compare/gatsby-plugin-cxs@2.0.6...gatsby-plugin-cxs@2.0.7) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-plugin-cxs
+
 ## [2.0.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-cxs/compare/gatsby-plugin-cxs@2.0.5...gatsby-plugin-cxs@2.0.6) (2019-02-19)
 
 **Note:** Version bump only for package gatsby-plugin-cxs